fix: requisito 6 correção

+++ desafio1.sql
@@ -5,11 +5,11 @@
 CREATE TABLE plans (
     plan_id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     plan_name VARCHAR(20) NOT NULL,
-    price DOUBLE NOT NULL
+    price DECIMAL(3,2) NOT NULL
 )  ENGINE=INNODB;
 
    INSERT INTO plans(plan_name, price)
-	VALUES("gratuito", 0.00),
+	VALUES("gratuito", 0),
 		("universitario", 5.99),
         ("pessoal", 6.99),
         ("familiar",7.99);
